backend/requirements.txt
coverage==4.5.3
Django==2.2.2
django-cors-headers==3.0.2
django-phone-field==1.7.1
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
